pyproject.toml
[project]
name = "hackeragenda"
version = "0.1.0"
description = "[Hackeragenda](http://hackeragenda.be) is a website that aggregate events in Belgium that are interesting for hackers of the hackerspace community. It does that by fetching data from other website (by scraping most of the time ...) *instead* of asking users to submit events on it. This is my modest attemp to answer the problem \"everyone is doing shitload of things and no one is aware of that and that sucks\"."
authors = [
    {name = "Laurent Peuch", email = "cortex@worlddomination.be"},
]
dependencies = [
    "asgiref~=3.6",
    "beautifulsoup4==4.9.1",
    "certifi==2020.6.20",
    "chardet==3.0.4",
    "dateparser==0.7.6",
    "django~=4.2",
    "django-debug-toolbar~=5.2",
    "django-extensions==3.0.3",
    "django-hamlpy~=1.4",
    "django-leaflet~=0.32",
    "django-pdb==0.6.2",
    "django-taggit~=6.1",
    "django-tastypie~=0.15",
    "django-view-dispatch==0.1",
    "facepy==1.0.12",
    "feedparser==6.0.*",
    "geographiclib==1.50",
    "geopy==2.0.0",
    "gunicorn==20.0.4",
    "hamlpy~=0.82",
    "html5lib==1.1",
    "icalendar==4.0.6",
    "idna==2.10",
    "markdown~=3.8",
    "mimeparse==0.1.3",
    "oauthlib==3.1.0",
    "Pygments==2.6.1",
    "python-dateutil==2.8.1",
    "python-mimeparse==1.6.0",
    "pytz==2020.1",
    "PyYAML==5.3.1",
    "regex==2020.7.14",
    "requests~=2.32",
    "requests-oauthlib==1.3.0",
    "six~=1.17",
    "soupsieve==2.0.1",
    "sqlparse==0.3.1",
    "tweepy==3.9.0",
    "tzlocal==2.1",
    "urllib3~=2.5",
    "webencodings==0.5.1",
    "weirdict==0.3.0",
    "Werkzeug==1.0.1",
    "legacy-cgi~=2.6",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "GPLV3+"}


[tool.pdm]
distribution = false

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
